feat(psffpp-upload): Replaced psffpp-upload with psffpp-upload2
