+++ src/commands/psffpp-upload.js
@@ -4,19 +4,22 @@
 
 // Public NPM libraries
 const axios = require('axios')
+const FormData = require('form-data')
+const fs = require('fs')
 
 // Local libraries
 const WalletUtil = require('../lib/wallet-util')
 
 const { Command, flags } = require('@oclif/command')
 
-class IpfsUpload extends Command {
+class IpfsUpload2 extends Command {
   constructor (argv, config) {
     super(argv, config)
 
     // Encapsulate dependencies.
     this.axios = axios
     this.walletUtil = new WalletUtil()
+    this.fs = fs
 
     // Bind 'this' object to all subfunctions.
     this.run = this.run.bind(this)
@@ -25,7 +28,7 @@ class IpfsUpload extends Command {
 
   async run () {
     try {
-      const { flags } = this.parse(IpfsUpload)
+      const { flags } = this.parse(IpfsUpload2)
 
       // Validate input flags
       this.validateFlags(flags)
@@ -36,7 +39,8 @@ class IpfsUpload extends Command {
       let server = this.walletUtil.getPsffppClient()
       server = server.psffppURL
 
-      await this.uploadFile({ path, fileName, server })
+      const result = await this.uploadFile({ path, fileName, server })
+      console.log('result: ', result)
 
       return true
     } catch (err) {
@@ -50,11 +54,17 @@ class IpfsUpload extends Command {
     try {
       const { path, fileName, server } = inObj
 
-      const result = await this.axios.post(`${server}/ipfs/upload`, {
-        path,
-        fileName
-      })
-      // console.log(`upload result: ${JSON.stringify(result.data, null, 2)}`)
+      const filePath = `${path}/${fileName}`
+
+      // Create a web form and append the file to it.
+      const form = new FormData()
+      const axiosConfig = {
+        headers: form.getHeaders()
+      }
+      form.append('file', this.fs.createReadStream(filePath), fileName)
+
+      // Send the file to the ipfs-file-stage server.
+      const result = await this.axios.post(`${server}/ipfs/upload`, form, axiosConfig)
 
       return result.data
     } catch (err) {
@@ -74,21 +84,16 @@ class IpfsUpload extends Command {
   }
 }
 
-IpfsUpload.description = `Upload a file to the IPFS node
+IpfsUpload2.description = `Upload a file to the psffpp-client
 
-If you are trying to pin a file to the PSFFPP network, you do not need to run
-this command. This command is automatically  used by the psffpp-pin command to
-upload the file.
-
-This command is useful in isolation if you want to test uploading and passing
-files between IPFS nodes, independant of the PSFFPP.
+This command will upload a file to the psffpp-client via the REST API.
 `
 
-IpfsUpload.flags = {
+IpfsUpload2.flags = {
   fileName: flags.string({
     char: 'f',
     description: 'filename to upload'
   })
 }
 
-module.exports = IpfsUpload
+module.exports = IpfsUpload2
